+++ apps/decodex/src/agent/app_server/tests.rs
@@ -25,11 +25,7 @@ use crate::{
 			REQUEST_TIMEOUT, RunRecorder, RuntimeConfigSummary, SkillsListResponse,
 			archive_app_server_thread_after_success, execute_app_server_run,
 			handle_dynamic_tool_call, handle_turn_execution_notification,
-			mcp_preflight_can_degrade, plugin_list_params_for_preflight, preflight_request,
-			preflight_request_with_timeout_retry, record_config_preflight,
-			record_interactive_request_state, record_mcp_preflight, record_mcp_preflight_degraded,
-			record_model_preflight, record_model_provider_preflight, record_plugin_preflight,
-			record_server_request, record_skills_preflight,
+			record_interactive_request_state, record_server_request,
 			record_thread_archive_result_best_effort, thread_archive_error_allows_discard,
 			validate_command_exec_health_check_result,
 		},
