+++ apps/decodex/src/orchestrator/tests/operator/status/http/state_endpoint.rs
@@ -1,321 +1,4 @@
-use std::panic;
-
-use crate::orchestrator::tests::operator::status::http::{
-	self, Arc, DashboardEventHub, Duration, ErrorKind, Mutex, OffsetDateTime,
-	OperatorControlRequests, PublishedOperatorSnapshot, Read as _, Shutdown, StateStore,
-	TcpListener, TcpStream, TempDir, TestEnvVarGuard, Value, Write, orchestrator, runtime, thread,
-};
-#[test]
-fn operator_state_endpoint_reads_complete_headers_before_parsing() {
-	let listener = TcpListener::bind("127.0.0.1:0").expect("listener should bind");
-	let address = listener.local_addr().expect("listener address should resolve");
-	let snapshot = Arc::new(Mutex::new(PublishedOperatorSnapshot::default()));
-	let state_store = Arc::new(StateStore::open_in_memory().expect("state store should open"));
-	let server_snapshot = Arc::clone(&snapshot);
-	let server_state_store = Arc::clone(&state_store);
-	let server = thread::spawn(move || {
-		let (stream, _) = listener.accept().expect("listener should accept a connection");
-		let dashboard_events = DashboardEventHub::default();
-
-		orchestrator::handle_operator_state_endpoint_connection(
-			stream,
-			&server_snapshot,
-			&dashboard_events,
-			&OperatorControlRequests::default(),
-			&server_state_store,
-		)
-		.expect("handler should accept segmented headers");
-	});
-	let mut client = TcpStream::connect(address).expect("client should connect");
-	let mut response = String::new();
-
-	client.write_all(b"GET /dash").expect("client should write first request fragment");
-
-	thread::sleep(Duration::from_millis(10));
-
-	client
-		.write_all(b"board HTTP/1.1\r\nHost: localhost\r\n\r\n")
-		.expect("client should write second request fragment");
-	client.shutdown(Shutdown::Write).expect("client should close the request body stream");
-	client.read_to_string(&mut response).expect("client should read response");
-	server.join().expect("server thread should complete");
-
-	assert!(response.starts_with("HTTP/1.1 200 OK\r\n"));
-	assert!(response.contains("<title>Decodex</title>"));
-}
-
-#[test]
-fn operator_state_endpoint_overlays_live_account_control_on_published_snapshot() {
-	const SNAPSHOT_UNIX_EPOCH: i64 = 1_774_000_000;
-
-	let temp_dir = TempDir::new().expect("temp dir should exist");
-	let _home_guard =
-		TestEnvVarGuard::set("HOME", temp_dir.path().to_str().expect("temp path should be UTF-8"));
-	let stale_snapshot = serde_json::json!({
-		"project_id": "all",
-		"account_control": {
-			"mode": "balanced",
-			"account_selector": null,
-		},
-		"accounts": [],
-		"current_lanes": [],
-		"recent_runs": [],
-		"history_lanes": [],
-		"queued_candidates": [],
-		"worktrees": [],
-		"post_review_lanes": [],
-	});
-
-	runtime::write_global_fixed_account_selector(Some("copy@example.com"))
-		.expect("global account selector should write");
-
-	let listener = TcpListener::bind("127.0.0.1:0").expect("listener should bind");
-	let address = listener.local_addr().expect("listener address should resolve");
-	let snapshot = Arc::new(Mutex::new(PublishedOperatorSnapshot {
-		snapshot_json: Some(
-			serde_json::to_vec(&stale_snapshot).expect("snapshot should serialize"),
-		),
-		last_publish_unix_epoch: Some(SNAPSHOT_UNIX_EPOCH),
-	}));
-	let state_store = Arc::new(StateStore::open_in_memory().expect("state store should open"));
-	let dashboard_events = DashboardEventHub::default();
-	let server_snapshot = Arc::clone(&snapshot);
-	let server_state_store = Arc::clone(&state_store);
-	let server_dashboard_events = dashboard_events.clone();
-	let server = thread::spawn(move || {
-		let (stream, _) = listener.accept().expect("listener should accept a connection");
-
-		orchestrator::handle_operator_state_endpoint_connection(
-			stream,
-			&server_snapshot,
-			&server_dashboard_events,
-			&OperatorControlRequests::default(),
-			&server_state_store,
-		)
-		.expect("handler should serve websocket snapshot");
-	});
-	let (mut client, response, mut frame) = http::open_dashboard_websocket_client(address);
-	let served_snapshot = http::read_websocket_json_until(&mut client, &mut frame, |payload| {
-		payload["type"] == "snapshot"
-	});
-
-	assert!(response.starts_with("HTTP/1.1 101 Switching Protocols\r\n"));
-	assert_eq!(served_snapshot["payload"]["snapshot"]["account_control"]["mode"], "fixed");
-	assert_eq!(
-		served_snapshot["payload"]["snapshot"]["account_control"]["account_selector"],
-		"copy@example.com"
-	);
-	assert_eq!(served_snapshot["payload"]["snapshot"]["project_id"], "all");
-	assert_eq!(served_snapshot["payload"]["snapshotPublishedAtUnixEpoch"], SNAPSHOT_UNIX_EPOCH);
-
-	drop(client);
-
-	dashboard_events.close_clients_for_test();
-	server.join().expect("server thread should complete");
-}
-
-#[test]
-fn operator_state_endpoint_serves_large_app_snapshot_without_truncation() {
-	const SNAPSHOT_UNIX_EPOCH: i64 = 1_774_000_000;
-
-	let temp_dir = TempDir::new().expect("temp dir should exist");
-	let _home_guard =
-		TestEnvVarGuard::set("HOME", temp_dir.path().to_str().expect("temp path should be UTF-8"));
-	let large_title = "x".repeat(2_000_000);
-	let snapshot_value = serde_json::json!({
-		"project_id": "all",
-		"account_control": {
-			"mode": "balanced",
-			"account_selector": null,
-		},
-		"accounts": [],
-		"current_lanes": [],
-		"recent_runs": [],
-		"history_lanes": [{
-			"issue_identifier": "PUB-100",
-			"title": large_title,
-		}],
-		"queued_candidates": [],
-		"worktrees": [],
-		"post_review_lanes": [],
-	});
-	let listener = TcpListener::bind("127.0.0.1:0").expect("listener should bind");
-
-	listener.set_nonblocking(true).expect("listener should be nonblocking");
-
-	let address = listener.local_addr().expect("listener address should resolve");
-	let snapshot = Arc::new(Mutex::new(PublishedOperatorSnapshot {
-		snapshot_json: Some(
-			serde_json::to_vec(&snapshot_value).expect("snapshot should serialize"),
-		),
-		last_publish_unix_epoch: Some(SNAPSHOT_UNIX_EPOCH),
-	}));
-	let state_store = Arc::new(StateStore::open_in_memory().expect("state store should open"));
-	let dashboard_events = DashboardEventHub::default();
-	let server_snapshot = Arc::clone(&snapshot);
-	let server_state_store = Arc::clone(&state_store);
-	let server_dashboard_events = dashboard_events.clone();
-	let server = thread::spawn(move || {
-		let stream = loop {
-			match listener.accept() {
-				Ok((stream, _)) => break stream,
-				Err(error) if error.kind() == ErrorKind::WouldBlock => {
-					thread::sleep(Duration::from_millis(5));
-				},
-				Err(error) => http::panic!("listener should accept a connection: {error}"),
-			}
-		};
-
-		orchestrator::handle_operator_state_endpoint_connection(
-			stream,
-			&server_snapshot,
-			&server_dashboard_events,
-			&OperatorControlRequests::default(),
-			&server_state_store,
-		)
-		.expect("handler should serve the complete large app snapshot");
-	});
-	let mut client = TcpStream::connect(address).expect("client should connect");
-	let mut response = Vec::new();
-
-	client
-		.write_all(
-			format!(
-				"GET {} HTTP/1.1\r\nHost: localhost\r\nConnection: close\r\n\r\n",
-				orchestrator::OPERATOR_APP_SNAPSHOT_ENDPOINT_PATH
-			)
-			.as_bytes(),
-		)
-		.expect("client should write request");
-	client.shutdown(Shutdown::Write).expect("client should close request body");
-
-	thread::sleep(Duration::from_millis(350));
-
-	client.read_to_end(&mut response).expect("client should read response");
-	server.join().expect("server thread should complete");
-
-	let header_end = response
-		.windows(4)
-		.position(|window| window == b"\r\n\r\n")
-		.expect("response should contain header terminator");
-	let headers = String::from_utf8(response[..header_end].to_vec())
-		.expect("response headers should be utf-8");
-	let content_length = headers
-		.lines()
-		.find_map(|line| line.strip_prefix("Content-Length: "))
-		.expect("response should contain Content-Length")
-		.parse::<usize>()
-		.expect("Content-Length should parse");
-	let body = &response[header_end + 4..];
-	let body_json: Value = serde_json::from_slice(body).expect("body should be complete JSON");
-
-	assert!(headers.starts_with("HTTP/1.1 200 OK\r\n"));
-	assert_eq!(body.len(), content_length);
-	assert_eq!(body_json["history_lanes"][0]["issue_identifier"], "PUB-100");
-	assert_eq!(
-		body_json["snapshotPublishedAtUnixEpoch"],
-		Value::Null,
-		"published epoch remains an HTTP header for app snapshot responses"
-	);
-}
-
-#[test]
-fn operator_state_endpoint_livez_ignores_poisoned_snapshot_lock() {
-	let listener = TcpListener::bind("127.0.0.1:0").expect("listener should bind");
-	let address = listener.local_addr().expect("listener address should resolve");
-	let snapshot = Arc::new(Mutex::new(PublishedOperatorSnapshot {
-		snapshot_json: Some(br#"{"status":"ok"}"#.to_vec()),
-		last_publish_unix_epoch: Some(OffsetDateTime::now_utc().unix_timestamp()),
-	}));
-	let state_store = Arc::new(StateStore::open_in_memory().expect("state store should open"));
-	let poisoned_snapshot = Arc::clone(&snapshot);
-	let _ = panic::catch_unwind(move || {
-		let _guard = poisoned_snapshot.lock().expect("snapshot lock should acquire");
-
-		http::panic!("poison snapshot lock");
-	});
-	let server_snapshot = Arc::clone(&snapshot);
-	let server_state_store = Arc::clone(&state_store);
-	let server = thread::spawn(move || {
-		let (stream, _) = listener.accept().expect("listener should accept a connection");
-		let dashboard_events = DashboardEventHub::default();
-
-		orchestrator::handle_operator_state_endpoint_connection(
-			stream,
-			&server_snapshot,
-			&dashboard_events,
-			&OperatorControlRequests::default(),
-			&server_state_store,
-		)
-		.expect("live probe should not require snapshot lock");
-	});
-	let mut client = TcpStream::connect(address).expect("client should connect");
-	let mut response = String::new();
-
-	client
-		.write_all(
-			format!(
-				"GET {} HTTP/1.1\r\nHost: localhost\r\nConnection: close\r\n\r\n",
-				orchestrator::OPERATOR_LIVE_ENDPOINT_PATH
-			)
-			.as_bytes(),
-		)
-		.expect("client should write request");
-	client.shutdown(Shutdown::Write).expect("client should close the request body stream");
-	client.read_to_string(&mut response).expect("client should read response");
-	server.join().expect("server thread should complete");
-
-	assert!(response.starts_with("HTTP/1.1 200 OK\r\n"));
-	assert!(response.ends_with("ok"));
-}
-
-#[test]
-fn operator_state_endpoint_serves_only_liveness_probe() {
-	let live_response = String::from_utf8(
-		orchestrator::build_operator_state_http_response(
-			format!(
-				"GET {} HTTP/1.1\r\nHost: localhost\r\nConnection: close\r\n\r\n",
-				orchestrator::OPERATOR_LIVE_ENDPOINT_PATH
-			)
-			.as_bytes(),
-		)
-		.expect("live response should build"),
-	)
-	.expect("live response should be utf-8");
-
-	assert!(live_response.starts_with("HTTP/1.1 200 OK\r\n"));
-	assert!(live_response.ends_with("ok"));
-}
-
-#[test]
-fn operator_state_endpoint_queues_linear_scan_request() {
-	let control_requests = OperatorControlRequests::default();
-	let body = br#"{"projectId":"pubfi"}"#;
-	let request = format!(
-		"POST {} HTTP/1.1\r\nHost: localhost\r\nContent-Type: application/json\r\nContent-Length: {}\r\nConnection: close\r\n\r\n{}",
-		orchestrator::OPERATOR_LINEAR_SCAN_ENDPOINT_PATH,
-		body.len(),
-		String::from_utf8_lossy(body)
-	);
-	let response = String::from_utf8(
-		orchestrator::build_operator_state_http_response_with_control_requests(
-			request.as_bytes(),
-			&control_requests,
-		)
-		.expect("linear scan response should build"),
-	)
-	.expect("linear scan response should be utf-8");
-	let body = response
-		.split_once("\r\n\r\n")
-		.map(|(_, body)| body)
-		.expect("linear scan response should include body");
-	let data: Value = serde_json::from_str(body).expect("linear scan response should be json");
-
-	assert!(response.starts_with("HTTP/1.1 202 Accepted\r\n"));
-	assert_eq!(data["status"], "queued");
-	assert_eq!(data["scope"], "pubfi");
-	assert_eq!(
-		control_requests.drain_linear_scan_requests().expect("linear scan requests should drain"),
-		vec![orchestrator::OperatorLinearScanRequest { project_id: Some(String::from("pubfi")) }]
-	);
-}
+mod http_request;
+mod linear_scan;
+mod live_account;
+mod liveness;
