+++ apps/decodex/src/orchestrator/tests/review_landing_orchestration/request_refresh.rs
@@ -1,317 +1,3 @@
-use crate::{
-	orchestrator::{
-		self, ReviewOrchestrationMarker, StateStore, TERMINAL_GUARDED_RUN_STATUS,
-		tests::{
-			self, FakePullRequestReviewStateInspector, FakeTracker,
-			TEST_EXTERNAL_REVIEW_REQUEST_COMMENT_ID, review_landing_status_support,
-		},
-	},
-	test_support,
-};
-
-#[test]
-fn reconcile_post_review_orchestration_requests_external_review_without_thumbs_up_baseline() {
-	let (temp_dir, config, workflow) = tests::temp_project_layout();
-	let config = tests::service_config_with_github_token_env_var(&config, "PATH");
-	let _path_guard = tests::install_fake_post_issue_comment_gh_response(
-		&temp_dir,
-		TEST_EXTERNAL_REVIEW_REQUEST_COMMENT_ID,
-		"2025-11-03T00:00:00Z",
-	);
-	let repo_root = config.repo_root().to_path_buf();
-	let issue = review_landing_status_support::post_review_sample_service_owned_issue("In Review");
-	let tracker =
-		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let head_oid = String::from_utf8(
-		test_support::hermetic_git_command()
-			.arg("-C")
-			.arg(&repo_root)
-			.args(["rev-parse", "HEAD"])
-			.output()
-			.expect("git rev-parse should run")
-			.stdout,
-	)
-	.expect("git output should be utf-8")
-	.trim()
-	.to_owned();
-	let pr_url = "https://github.com/hack-ink/decodex/pull/173";
-
-	state_store
-		.upsert_worktree("pubfi", &issue.id, "main", &repo_root.display().to_string())
-		.expect("worktree should record");
-
-	tests::seed_review_handoff_marker_for_path(
-		&state_store,
-		config.service_id(),
-		&repo_root,
-		&tests::sample_review_handoff_marker("main", pr_url, &head_oid),
-	);
-	tests::seed_review_orchestration_marker_for_path(
-		&state_store,
-		config.service_id(),
-		&repo_root,
-		&ReviewOrchestrationMarker::new(
-			"run-1",
-			1,
-			"main",
-			pr_url,
-			&head_oid,
-			"request_pending",
-			None,
-			None,
-			None,
-			0,
-			0,
-			None,
-		),
-	);
-
-	let initial_review_state = tests::sample_pull_request_review_state(
-		pr_url,
-		"main",
-		&head_oid,
-		Some("APPROVED"),
-		"MERGEABLE",
-		"CLEAN",
-		Some("SUCCESS"),
-		0,
-	);
-
-	orchestrator::reconcile_post_review_orchestration_with_inspector(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		&FakePullRequestReviewStateInspector::new(vec![Ok(initial_review_state)]),
-	)
-	.expect("post-review orchestration should succeed");
-
-	let marker = tests::persisted_review_orchestration_marker_for_path(
-		&state_store,
-		config.service_id(),
-		&repo_root,
-	);
-
-	assert_eq!(marker.phase(), "waiting_for_ack");
-	assert_eq!(marker.request_description_thumbs_up_count(), None);
-}
-
-#[test]
-fn reconcile_post_review_orchestration_filters_terminal_identifier_worktree_before_tracker_refresh()
-{
-	let (_temp_dir, config, workflow) = tests::temp_project_layout();
-	let repo_root = config.repo_root().to_path_buf();
-	let issue = review_landing_status_support::post_review_sample_service_owned_issue("Todo");
-	let tracker = FakeTracker::new(vec![issue.clone()]);
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let stale_issue_id = "PUB-001";
-	let stale_worktree_path = config.worktree_root().join(stale_issue_id);
-
-	state_store
-		.record_run_attempt("run-01", stale_issue_id, 1, TERMINAL_GUARDED_RUN_STATUS)
-		.expect("terminal run attempt should record");
-	state_store
-		.upsert_worktree("pubfi", &issue.id, "main", &repo_root.display().to_string())
-		.expect("valid worktree should record");
-	state_store
-		.upsert_worktree(
-			"pubfi",
-			stale_issue_id,
-			"x/pubfi-pub-001",
-			&stale_worktree_path.display().to_string(),
-		)
-		.expect("stale worktree mapping should record");
-
-	orchestrator::reconcile_post_review_orchestration_with_inspector(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		&FakePullRequestReviewStateInspector::new(Vec::new()),
-	)
-	.expect("post-review orchestration should ignore stale terminal local residue");
-
-	let refresh_queries = tracker.refresh_queries.borrow();
-
-	assert!(
-		refresh_queries.iter().flatten().any(|issue_id| issue_id == &issue.id),
-		"valid worktree issue id should still be sent to tracker refresh"
-	);
-	assert!(
-		refresh_queries.iter().flatten().all(|issue_id| issue_id != stale_issue_id),
-		"terminal local identifier residue must not be sent to post-review tracker refresh"
-	);
-}
-
-#[test]
-fn reconcile_post_review_orchestration_uses_matching_handoff_record_for_current_branch() {
-	let (temp_dir, config, workflow) = tests::temp_project_layout();
-	let config = tests::service_config_with_github_token_env_var(&config, "PATH");
-	let _path_guard = tests::install_fake_post_issue_comment_gh_response(
-		&temp_dir,
-		TEST_EXTERNAL_REVIEW_REQUEST_COMMENT_ID,
-		"2025-11-03T00:00:00Z",
-	);
-	let repo_root = config.repo_root().to_path_buf();
-	let issue = review_landing_status_support::post_review_sample_service_owned_issue("In Review");
-	let tracker =
-		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let head_oid = String::from_utf8(
-		test_support::hermetic_git_command()
-			.arg("-C")
-			.arg(&repo_root)
-			.args(["rev-parse", "HEAD"])
-			.output()
-			.expect("git rev-parse should run")
-			.stdout,
-	)
-	.expect("git output should be utf-8")
-	.trim()
-	.to_owned();
-	let pr_url = "https://github.com/hack-ink/decodex/pull/173";
-	let current_branch = "main";
-
-	state_store
-		.upsert_worktree(
-			config.service_id(),
-			&issue.id,
-			current_branch,
-			&repo_root.display().to_string(),
-		)
-		.expect("worktree should record");
-
-	tests::seed_review_handoff_marker_for_path(
-		&state_store,
-		config.service_id(),
-		&repo_root,
-		&tests::sample_review_handoff_marker(current_branch, pr_url, &head_oid),
-	);
-	tests::seed_review_orchestration_marker_for_path(
-		&state_store,
-		config.service_id(),
-		&repo_root,
-		&ReviewOrchestrationMarker::new(
-			"run-1",
-			1,
-			current_branch,
-			pr_url,
-			&head_oid,
-			"request_pending",
-			None,
-			None,
-			None,
-			0,
-			0,
-			None,
-		),
-	);
-
-	let review_state = tests::sample_pull_request_review_state(
-		pr_url,
-		current_branch,
-		&head_oid,
-		Some("APPROVED"),
-		"MERGEABLE",
-		"CLEAN",
-		Some("SUCCESS"),
-		0,
-	);
-
-	orchestrator::reconcile_post_review_orchestration_with_inspector(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		&FakePullRequestReviewStateInspector::new(vec![Ok(review_state)]),
-	)
-	.expect("post-review orchestration should succeed");
-
-	let marker = tests::persisted_review_orchestration_marker_for_path(
-		&state_store,
-		config.service_id(),
-		&repo_root,
-	);
-
-	assert_eq!(marker.phase(), "waiting_for_ack");
-	assert_eq!(marker.pr_url(), pr_url);
-}
-
-#[test]
-fn reconcile_post_review_orchestration_rebinds_stale_head_marker_after_repair_push() {
-	let (temp_dir, config, workflow) = tests::temp_project_layout();
-	let config = tests::service_config_with_github_token_env_var(&config, "PATH");
-	let _path_guard = tests::install_fake_post_issue_comment_gh_response(
-		&temp_dir,
-		TEST_EXTERNAL_REVIEW_REQUEST_COMMENT_ID,
-		"2025-11-03T00:00:00Z",
-	);
-	let repo_root = config.repo_root().to_path_buf();
-	let issue = review_landing_status_support::post_review_sample_service_owned_issue("In Review");
-	let tracker =
-		FakeTracker::with_refresh_snapshots(vec![issue.clone()], vec![vec![issue.clone()]]);
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let pr_url = "https://github.com/hack-ink/decodex/pull/173";
-	let marker_head_oid = tests::git_output(&repo_root, &["rev-parse", "HEAD"]);
-	let current_head_oid =
-		tests::commit_worktree_change(&repo_root, "repair.txt", "repair push\n", "repair push");
-
-	state_store
-		.upsert_worktree("pubfi", &issue.id, "main", &repo_root.display().to_string())
-		.expect("worktree should record");
-
-	tests::seed_review_handoff_marker_for_path(
-		&state_store,
-		config.service_id(),
-		&repo_root,
-		&tests::sample_review_handoff_marker("main", pr_url, &marker_head_oid),
-	);
-	tests::seed_review_orchestration_marker_for_path(
-		&state_store,
-		config.service_id(),
-		&repo_root,
-		&tests::sample_review_orchestration_marker(
-			"main",
-			pr_url,
-			&marker_head_oid,
-			"waiting_for_result",
-			1,
-		),
-	);
-
-	let review_state = tests::sample_pull_request_review_state(
-		pr_url,
-		"main",
-		&current_head_oid,
-		Some("APPROVED"),
-		"MERGEABLE",
-		"CLEAN",
-		Some("SUCCESS"),
-		0,
-	);
-
-	orchestrator::reconcile_post_review_orchestration_with_inspector(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		&FakePullRequestReviewStateInspector::new(vec![Ok(review_state)]),
-	)
-	.expect("post-review orchestration should rebind stale marker without attention");
-
-	let marker = tests::persisted_review_orchestration_marker_for_path(
-		&state_store,
-		config.service_id(),
-		&repo_root,
-	);
-
-	assert_eq!(marker.phase(), "waiting_for_ack");
-	assert_eq!(marker.head_sha(), current_head_oid);
-	assert_eq!(marker.request_comment_database_id(), Some(TEST_EXTERNAL_REVIEW_REQUEST_COMMENT_ID));
-	assert_eq!(marker.external_round_count(), 1);
-	assert!(tracker.comments.borrow().is_empty());
-	assert!(tracker.state_updates.borrow().is_empty());
-	assert!(tracker.label_additions.borrow().is_empty());
-	assert!(tracker.label_removals.borrow().is_empty());
-}
+mod external_request;
+mod stale_repair;
+mod tracker_refresh;
